eCAL v5.6.0: * thirdparty:   * new dependencies (mainly required by the new ecal recorder)     * curl     * fineftp-server     * hdf5     * libssh2     * tinyxml2     * zlib * core:   * bypass local firewalls or vpn by broadcasting monitoring on local host * apps:   * ecal recorder     * major redesign to support new features like ..     * measurement history     * delete measurements     * add comments to measurements     * further information about connected recorder clients     * save configuration     * upload (merge) distributed measurements to the host machine     * upload (merge) distributed to any custom FTP server     * rpc interface to remote-control the recorder GUI     * warn in case of faulty configuration     * improved command line interface     * addon support to integrate third party recording clients   * ecal player     * edit labels

+++ app/rec/rec_gui/src/widgets/config_widget/config_widget.h
@@ -34,7 +34,7 @@ private slots:
   void measurementRootDirectoryChanged(const std::string& root_dir);
   void measurementNameChanged         (const std::string& name);
   void updatePathPreview              ();
-  void maxFileSizeChanged             (size_t megabytes);
+  void maxFileSizeChanged             (unsigned int megabytes);
   void descriptionChanged             (const std::string& description);
 
 private:
